+++ app/src/main/java/com/gdsciiita/ontimepro/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.on_time_pro;
+package com.gdsciiita.ontimepro;
 
 import androidx.appcompat.app.AppCompatActivity;
