+++ app/src/main/java/com/gdsciiita/ontimepro/ProfileActivity.java
@@ -1,4 +1,4 @@
-package com.example.on_time_pro;
+package com.gdsciiita.ontimepro;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
