+++ app/src/main/java/com/gdsciiita/ontimepro/fragments/MainFragment.kt
@@ -1,11 +1,11 @@
-package com.example.on_time_pro.fragments
+package com.gdsciiita.ontimepro.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.on_time_pro.R
+import com.gdsciiita.ontimepro.R
 
 class MainFragment : Fragment() {
