+++ app/src/main/java/com/gdsciiita/ontimepro/fragments/ScheduleFragment.kt
@@ -1,20 +1,20 @@
-package com.example.on_time_pro.fragments
+package com.gdsciiita.ontimepro.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.on_time_pro.R
+import com.gdsciiita.ontimepro.R
 
-class EmailAttachFragment : Fragment() {
+class ScheduleFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_email_attach, container, false)
+        return inflater.inflate(R.layout.fragment_schedule, container, false)
     }
 
 }
